test(attribute): group default-value and is_default tests in their own file
Move the is_default preservation test and the inherited default-backed
materialisation tests out of the large test_attribute module into a dedicated
test_attribute_default_value module, so the tests that share the default-value
/ is_default concern live together.

All inherited cases build their node through a single setup helper that returns
the node id and the schema before and after inheritance. The reproduction test
reuses that setup, then runs the real add-attribute migration and the save,
exercising the whole chain in one test without duplicating the schema evolution.
